+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * user can set up which task list will be used both in the notebook parameters and in the Tools menu
 * tasks status sync (if you mark the task on the server, it will propagate to the notebook)
 * Zim start date pre-filled
-* explicitly download all non-complete history 
+* explicitly download all non-complete history
+* relative postponing buttons  
 
 # 1.0 (2018-06-18) 0.68 compatible
 Fully working version, last compatible with Zim 0.68 and older. Current master supports only newer versions.